Move to stringExpressions and address comments.
